pyproject.toml
[project]
name = "jit-mcp"
version = "0.1.0"
description = "Just-In-Time Model Context Protocol (JIT-MCP) agentic system"
readme = "README.md"
authors = [
    { name = "Karthik Rajasekaran", email = "krajasek@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "mcp>=1.2.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
